Add gnu.cfg library. Make Cppcheck work without system headers (faster). - gnu.cfg replaces __typeof__ with typeof() and triggered some misra-c2012-10.4 - fixed by making both function args the same essential type

+++ board/safety/safety_hyundai_canfd.h
@@ -127,8 +127,8 @@ RxCheck hyundai_canfd_hda2_long_rx_checks[] = {
 
 
 
-const int HYUNDAI_PARAM_CANFD_ALT_BUTTONS = 32;
-const int HYUNDAI_PARAM_CANFD_HDA2_ALT_STEERING = 128;
+const uint16_t HYUNDAI_PARAM_CANFD_ALT_BUTTONS = 32;
+const uint16_t HYUNDAI_PARAM_CANFD_HDA2_ALT_STEERING = 128;
 bool hyundai_canfd_alt_buttons = false;
 bool hyundai_canfd_hda2_alt_steering = false;
